Add files via upload
Upon further investigation, there are a lot of problems going on.  Firstly, I put in a threshold of three because when the input was "Nine" that looked like a good line of separation.  But with many other inputs, the real match has a spike value of 20 or 40  and every other correlation would spike to ~10, so a threshold of 3 wouldn't work at all.  However, I can't raise the threshold because then numbers like "nine" wouldn't find a match at all.  Additionally, some really mysterious stuff is going on with Two.  You can see in the 'Correlation against 'Two'' figure that even though word_in is set equal to two_ref, the correlation of two_ref against word-in is very different from two_ref and two_ref.  Doesn't make any sense.  I'm thinking to fix the bigger threshold issue, we should instead look for the spike nearest to the center, as that indicates that there is a similarity in signals that are aligned in time.  The complication here is that the spike is only in the 'center' when plotted against the time vector generated by the xcorr function.  When looking at the actual location of the spike in the _corr vector, it might be anywhere from index 12000 to 20480, so we would need a way to measure 'centerness' flexibly to compare which spike is most central.